cleanup and fixed pull.sh

+++ pull.sh
@@ -1,7 +1,11 @@
-git st
-git pull --rebase
-rm resume/*
-cp public/resume/* resume/
-git add -A
-git cm $1
-git push
+if [ -z "$1" ] && [ -z "$2" ]; then
+  echo "🌺 git commit message missing!"
+else
+  git st
+  git pull --rebase
+  rm resume/*
+  cp public/resume/* resume/
+  git add -A
+  git cm "$1"
+  git push
+fi
